requirements.txt
black==24.4.2
faiss-cpu==1.8.0
huggingface-hub==0.23.3
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-huggingface==0.0.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.18
numpy==1.26.4
openai==1.33.0
pydantic==2.7.3
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
sentence-transformers==3.0.1
transformers==4.41.2
yarl==1.9.4
fastapi==0.111.0
uvicorn[standard]==0.30.1